fix(linux): Session 34 shadow rendering and 64-bit fixes

+++ GeneralsMD/Code/GameEngineDevice/Source/W3DDevice/GameClient/W3DAssetManager.cpp
@@ -770,7 +770,8 @@ RenderObjClass * W3DAssetManager::Create_Render_Obj(
 		const char *mesh_name = strchr (name, '.');
 		if (mesh_name != nullptr)
 		{
-			lstrcpyn(filename, name, ((int)mesh_name) - ((int)name) + 1);
+			// TheSuperHackers @bugfix fighter19 13/02/2026 BenderAI - Use intptr_t for pointer arithmetic (64-bit compat)
+			lstrcpyn(filename, name, ((intptr_t)mesh_name) - ((intptr_t)name) + 1);
 			lstrcat(filename, ".w3d");
 		} else {
 			sprintf( filename, "%s.w3d", name);
@@ -1370,7 +1371,8 @@ RenderObjClass * W3DAssetManager::Create_Render_Obj(const char * name,float scal
 		char filename [MAX_PATH];
 		char *mesh_name = ::strchr (name, '.');
 		if (mesh_name != nullptr) {
-			::lstrcpyn (filename, name, ((int)mesh_name) - ((int)name) + 1);
+			// TheSuperHackers @bugfix fighter19 13/02/2026 BenderAI - Use intptr_t for pointer arithmetic (64-bit compat)
+			::lstrcpyn (filename, name, ((intptr_t)mesh_name) - ((intptr_t)name) + 1);
 			if (isGranny)
 				::lstrcat (filename, ".gr2");
 			else
